move tests to unit dir
