+++ test/unit/es-bib.test.js
@@ -1,12 +1,12 @@
 const expect = require('chai').expect
 
-const SierraBib = require('../lib/sierra-models/bib')
-const EsBib = require('../lib/es-models/bib')
+const SierraBib = require('../../lib/sierra-models/bib')
+const EsBib = require('../../lib/es-models/bib')
 
 describe('EsBib', function () {
   describe('constructor', function () {
     it('initializes an EsBib with a \'bib\' property', function () {
-      const record = new SierraBib(require('./fixtures/bib-10001936.json'))
+      const record = new SierraBib(require('../fixtures/bib-10001936.json'))
       const esBib = new EsBib(record)
       expect(esBib.bib).to.eq(record)
     })
@@ -14,14 +14,14 @@ describe('EsBib', function () {
 
   describe('uri', async function () {
     it('should return correctly prefixed uri for nypl bib', async function () {
-      const record = new SierraBib(require('./fixtures/bib-10001936.json'))
+      const record = new SierraBib(require('../fixtures/bib-10001936.json'))
       const esBib = new EsBib(record)
       const uri = await esBib.uri()
       expect(uri).to.eq('b10001936')
     })
 
     it('should return correctly prefixed uri for partner bib', async function () {
-      const record = new SierraBib(require('./fixtures/bib-hl990000453050203941.json'))
+      const record = new SierraBib(require('../fixtures/bib-hl990000453050203941.json'))
       const esBib = new EsBib(record)
       const uri = await esBib.uri()
       expect(uri).to.eq('hb990000453050203941')
@@ -30,7 +30,7 @@ describe('EsBib', function () {
 
   describe('title', function () {
     it('should return array of titles', function () {
-      const record = new SierraBib(require('./fixtures/bib-11606020.json'))
+      const record = new SierraBib(require('../fixtures/bib-11606020.json'))
       const esBib = new EsBib(record)
       expect(esBib.title()).to.deep.equal(
         ['880-02 Sefer Toldot Yeshu = The gospel according to the Jews, called Toldoth Jesu : the generations of Jesus, now first translated from the Hebrew.',
@@ -42,7 +42,7 @@ describe('EsBib', function () {
 
   describe('parallelTitle', function () {
     it('should return array of parallel titles', function () {
-      const record = new SierraBib(require('./fixtures/bib-11606020.json'))
+      const record = new SierraBib(require('../fixtures/bib-11606020.json'))
       const esBib = new EsBib(record)
       expect(esBib.parallelTitle()).to.deep.equal(
         [
@@ -55,7 +55,7 @@ describe('EsBib', function () {
 
   describe('creatorLiteral', function () {
     it('should return the creator literal', function () {
-      const record = new SierraBib(require('./fixtures/bib-10001936.json'))
+      const record = new SierraBib(require('../fixtures/bib-10001936.json'))
       const esBib = new EsBib(record)
       expect(esBib.creatorLiteral()).to.deep.equal(['Shermazanian, Galust.'])
     })
@@ -63,7 +63,7 @@ describe('EsBib', function () {
 
   describe('parallelCreatorLiteral', function () {
     it('should return the parallel creator literals', function () {
-      const record = new SierraBib(require('./fixtures/bib-hl-990137923810203941.json'))
+      const record = new SierraBib(require('../fixtures/bib-hl-990137923810203941.json'))
       const esBib = new EsBib(record)
       expect(esBib.parallelCreatorLiteral()).to.deep.equal(
         ['بوريني، حسن احمد.']
@@ -73,7 +73,7 @@ describe('EsBib', function () {
 
   describe('contentsTitle', function () {
     it('should return the contents title values in an array', function () {
-      const record = new SierraBib(require('./fixtures/bib-11055155.json'))
+      const record = new SierraBib(require('../fixtures/bib-11055155.json'))
       const esBib = new EsBib(record)
       expect(esBib.contentsTitle()).to.deep.equal(
         [
@@ -92,7 +92,7 @@ describe('EsBib', function () {
 
   describe('contributor_sort', function () {
     it('should return the first contributor literal, truncated to 80 characters and lower case', function () {
-      const record = new SierraBib(require('./fixtures/bib-hl990000453050203941.json'))
+      const record = new SierraBib(require('../fixtures/bib-hl990000453050203941.json'))
       const esBib = new EsBib(record)
       expect(esBib.contributor_sort()).to.equal('ginosar, sh. (shaleṿ), 1902- ed.')
     })
@@ -100,7 +100,7 @@ describe('EsBib', function () {
 
   describe('contributorLiteral', function () {
     it('should return the contributorLiteral', function () {
-      const record = new SierraBib(require('./fixtures/bib-hl990000453050203941.json'))
+      const record = new SierraBib(require('../fixtures/bib-hl990000453050203941.json'))
       const esBib = new EsBib(record)
       expect(esBib.contributorLiteral()).to.deep.equal(
         [
@@ -112,7 +112,7 @@ describe('EsBib', function () {
 
   describe('parallelContributorLiteral', function () {
     it('should return parallel contributor fields', function () {
-      const record = new SierraBib(require('./fixtures/bib-parallels-party.json'))
+      const record = new SierraBib(require('../fixtures/bib-parallels-party.json'))
       const esBib = new EsBib(record)
       expect(esBib.parallelContributorLiteral()).to.deep.equal(
         ['710-07/$1 parallel content for 710$a parallel content for 710$z']
@@ -122,7 +122,7 @@ describe('EsBib', function () {
 
   describe('numItems', function () {
     it('should return 0', function () {
-      const record = new SierraBib(require('./fixtures/bib-hl990000453050203941.json'))
+      const record = new SierraBib(require('../fixtures/bib-hl990000453050203941.json'))
       const esBib = new EsBib(record)
       expect(esBib.numItems()).to.equal(0)
     })
@@ -130,7 +130,7 @@ describe('EsBib', function () {
 
   describe('type', function () {
     it('should return \'nypl:Item\'', function () {
-      const record = new SierraBib(require('./fixtures/bib-hl990000453050203941.json'))
+      const record = new SierraBib(require('../fixtures/bib-hl990000453050203941.json'))
       const esBib = new EsBib(record)
       expect(esBib.type()).to.equal('nypl:Item')
     })
@@ -138,13 +138,13 @@ describe('EsBib', function () {
 
   describe('nyplSource', function () {
     it('should return the source for partner items', function () {
-      const record = new SierraBib(require('./fixtures/bib-hl990000453050203941.json'))
+      const record = new SierraBib(require('../fixtures/bib-hl990000453050203941.json'))
       const esBib = new EsBib(record)
       expect(esBib.nyplSource()).to.equal('recap-hl')
     })
 
     it('should return the source for NYPL items', function () {
-      const record = new SierraBib(require('./fixtures/bib-10001936.json'))
+      const record = new SierraBib(require('../fixtures/bib-10001936.json'))
       const esBib = new EsBib(record)
       expect(esBib.nyplSource()).to.equal('sierra-nypl')
     })
